int64 is the default type after parsing

+++ controlflow.goto.go
@@ -8,7 +8,7 @@ type InstructionControlFlowGoto struct {
 	Instruction
 }
 
-func (i *InstructionControlFlowGoto) Goto(index int) error {
+func (i *InstructionControlFlowGoto) Goto(index int64) error {
 	variable, found := i.Env.Get("SYS_CUR")
 	if !found {
 		return fmt.Errorf("invalid variable CUR")
